// resources/js/Pages/SPT/TablistBadan/L12A/types.ts
export interface L12AData {
    id?: string;
    spt_badan_id?: string;
    taxable_income: number;
    pph_payable: number;
    dpp: number;
    pph_26_a: boolean | null;
    pph_26_a_value: number;
    pph_26_b: boolean | null;
    pph_26_b_1: boolean | null;
    pph_26_b_1_value: number;
    pph_26_b_2: boolean | null;
    pph_26_b_2_a: boolean | null;
    pph_26_b_2_a_value: string | null;
    pph_26_b_2_b: boolean | null;
    pph_26_b_2_b_value: string | null;
    pph_26_b_2_c: boolean | null;
    pph_26_b_2_d: boolean | null;
}